refactored test setup
