+++ dart/async/pubspec.yaml
@@ -1,5 +1,5 @@
 name: pageloader
-version: 2.0.0
+version: 2.0.1
 author: Marc Fisher II <fisherii@google.com>
 description: >
   Supports the creation of page objects that can be shared between in-browser tests
