+++ dart/async/test/webdriver_no_shadow_dom_test.dart
@@ -12,65 +12,17 @@
 // limitations under the License.
 
 @TestOn('vm')
-library pageloader.test.webdriver;
+library pageloader.test.webdriver_no_shadow_dom;
 
-import 'dart:async';
-import 'dart:io';
-
-import 'package:pageloader/webdriver.dart' show WebDriverPageLoader;
-import 'package:path/path.dart' as path;
+import 'package:pageloader/webdriver.dart';
 import 'package:test/test.dart';
-import 'package:webdriver/io.dart' show Capabilities, WebDriver, createDriver;
+import 'package:webdriver/io.dart' show WebDriver;
 
-import 'src/common.dart' as plt;
-import 'src/shared.dart' as shared;
+import 'setup/webdriver_test_setup.dart' show runTests;
 
 void main() {
-  WebDriver driver;
-
-  setUp(() async {
-    driver = await _createTestDriver();
-    await driver.get(_testPagePath);
-    shared.loader = new WebDriverPageLoader(driver, useShadowDom: false);
-  });
-
-  tearDown(() async {
-    await driver.quit();
-    shared.loader = null;
-  });
-
-  plt.runTests();
+  runTests(pageLoaderFactory, 'webdriver_no_shadow_dom_test_page.html');
 }
 
-String get _testPagePath {
-  var testPagePath =
-      path.join('test', 'data', 'webdriver_no_shadow_dom_test_page.html');
-
-  testPagePath = path.absolute(testPagePath);
-  if (!FileSystemEntity.isFileSync(testPagePath)) {
-    throw new Exception('Could not find the test file at "$testPagePath".'
-        ' Make sure you are running tests from the root of the project.');
-  }
-  return path.toUri(testPagePath).toString();
-}
-
-Future<WebDriver> _createTestDriver() {
-  Map capabilities = Capabilities.chrome;
-  Map env = Platform.environment;
-
-  Map chromeOptions = {};
-
-  if (env['CHROMEDRIVER_BINARY'] != null) {
-    chromeOptions['binary'] = env['CHROMEDRIVER_BINARY'];
-  }
-
-  if (env['CHROMEDRIVER_ARGS'] != null) {
-    chromeOptions['args'] = env['CHROMEDRIVER_ARGS'].split(' ');
-  }
-
-  if (chromeOptions.isNotEmpty) {
-    capabilities['chromeOptions'] = chromeOptions;
-  }
-
-  return createDriver(desired: capabilities);
-}
+PageLoader pageLoaderFactory(WebDriver driver) =>
+    new WebDriverPageLoader(driver, useShadowDom: false);
